dedupe utils
